requirements.txt
opencv-python-headless
streamlit
streamlit-webrtc
mediapipe
gTTS
Pillow
numpy
av